Implemented AuthService in Components
