+++ lib/services/auth/auth_provider.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_course/services/auth/auth_user.dart';
 
 abstract class AuthProvider {
+  Future<void> initialize();
   AuthUser? get currentUser;
   Future<AuthUser> logIn({
     required String email,
