+++ lib/services/auth/auth_service.dart
@@ -1,10 +1,17 @@
 import 'package:flutter_course/services/auth/auth_provider.dart';
 import 'package:flutter_course/services/auth/auth_user.dart';
+import 'package:flutter_course/services/auth/firebase_auth_provider.dart';
 
 class AuthService implements AuthProvider {
   final AuthProvider provider;
+  const AuthService(this.provider);
 
-  AuthService(this.provider);
+  factory AuthService.firebase() => AuthService(
+        FirebaseAuthProvider(),
+      );
+
+  @override
+  Future<void> initialize() => provider.initialize();
 
   @override
   Future<AuthUser> createUser({
@@ -17,7 +24,6 @@ class AuthService implements AuthProvider {
       );
 
   @override
-  // TODO: implement currentUser
   AuthUser? get currentUser => provider.currentUser;
   @override
   Future<AuthUser> logIn({
